Refactor metric reporter module
