+++ ftests/metrics/src/main/resources/META-INF/beans.xml
@@ -14,6 +14,6 @@
        xsi:schemaLocation="http://xmlns.jcp.org/xml/ns/javaee http://xmlns.jcp.org/xml/ns/javaee/beans_1_1.xsd"
        version="1.1" bean-discovery-mode="all">
     <interceptors>
-        <class>org.commonjava.indy.subsys.metrics.jaxrs.interceptor.MetricsInterceptor</class>
+        <class>org.commonjava.indy.subsys.metrics.interceptor.IndyMetricsInterceptor</class>
     </interceptors>
 </beans>
