+++ subsys/metrics/core/src/main/conf/conf.d/elasticsearch.properties
