+++ subsys/metrics/core/src/main/resources/META-INF/beans.xml
@@ -2,6 +2,6 @@
 <beans xmlns="http://java.sun.com/xml/ns/javaee" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
        xsi:schemaLocation="http://java.sun.com/xml/ns/javaee http://java.sun.com/xml/ns/javaee/beans_1_0.xsd">
     <interceptors>
-        <class>org.commonjava.indy.subsys.metrics.jaxrs.interceptor.MetricsInterceptor</class>
+        <class>org.commonjava.indy.subsys.metrics.interceptor.IndyMetricsInterceptor</class>
     </interceptors>
 </beans>
