+++ subsys/metrics/core/src/main/resources/default-metrics.conf
