--- subsys/metrics/reporter/src/main/java/org/commonjava/indy/subsys/metrics/jaxrs/producer/IndyMetricProducer.java
@@ -1,76 +1,0 @@
-/**
- * Copyright (C) 2011-2020 Red Hat, Inc. (https://github.com/Commonjava/indy)
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *         http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package org.commonjava.indy.subsys.metrics.jaxrs.producer;
-
-import com.codahale.metrics.MetricRegistry;
-import com.codahale.metrics.health.HealthCheckRegistry;
-import org.commonjava.indy.action.IndyLifecycleException;
-import org.commonjava.indy.action.StartupAction;
-import org.commonjava.indy.subsys.metrics.IndyMetricsManager;
-
-import javax.enterprise.context.ApplicationScoped;
-import javax.enterprise.inject.Produces;
-import javax.inject.Inject;
-
-/**
- * Created by xiabai on 2/27/17.
- */
-public class IndyMetricProducer implements StartupAction
-{
-
-    private static final String METRIC_ID = "metric";
-
-    @ApplicationScoped
-    @Produces
-    public MetricRegistry getMetricRegistry()
-    {
-        return new MetricRegistry();
-    }
-
-    @ApplicationScoped
-    @Produces
-    public HealthCheckRegistry getHealthCheckRegistry() {
-        return new HealthCheckRegistry();
-    }
-
-    @Override
-    public String getId()
-    {
-        return METRIC_ID;
-    }
-
-    @Inject
-    private IndyMetricsManager indyMetricsManager;
-
-    @Override
-    public void start() throws IndyLifecycleException
-    {
-        try
-        {
-            indyMetricsManager.startReporter();
-        }
-        catch ( Exception e )
-        {
-            throw new IndyLifecycleException( "Failed while attempting to start Metric reporters.", e );
-        }
-    }
-
-    @Override
-    public int getStartupPriority()
-    {
-        return 10;
-    }
-}
